Refactor: Make everyone happy
