+++ development/coverage_plugins/tests/test_jinja.py
@@ -4,20 +4,17 @@
 from __future__ import annotations
 
 import importlib.util
-import os
 import shutil
-import subprocess
-import sys
-from pathlib import Path
 from typing import TYPE_CHECKING, cast
 
 import pytest
-from coverage import Coverage, CoverageData
+from coverage import Coverage
 from coverage.exceptions import ConfigError
 from coverage_plugins.jinja import JinjaTemplateCoveragePlugin, JinjaTemplateFileReporter, coverage_init
 from jinja2 import Environment, FileSystemLoader, ModuleLoader
 
 if TYPE_CHECKING:
+    from pathlib import Path
     from types import FrameType
 
     from coverage.results import Analysis
@@ -287,6 +284,103 @@ def test_reporter_translates_loop_backedge_to_no_else_endif_arc(tmp_path: Path) 
 
     assert reporter.arcs() == {(1, 2), (2, 3), (2, 4)}
     assert (2, 4) in reporter.translate_arcs([(2, 1)])
+
+
+@pytest.mark.parametrize(
+    ("source", "recorded_arc", "expected_arc"),
+    [
+        (
+            "{% block content %}\n{% for item in items %}\n{% if item.enabled %}\n{{ item.name }}\n{% endif %}\n{% endfor %}\n{% endblock %}\n",
+            (3, 2),
+            (3, 5),
+        ),
+        (
+            "{% call render_items() %}\n{% for item in items %}\n{% if item.enabled %}\n{{ item.name }}\n{% endif %}\n{% endfor %}\n{% endcall %}\n",
+            (3, 2),
+            (3, 5),
+        ),
+        (
+            "{% filter upper %}\n{% for item in items %}\n{% if item.enabled %}\n{{ item.name }}\n{% endif %}\n{% endfor %}\n{% endfilter %}\n",
+            (3, 2),
+            (3, 5),
+        ),
+        (
+            "{% with enabled = item.enabled %}\n{% for item in items %}\n{% if enabled %}\n{{ item.name }}\n{% endif %}\n{% endfor %}\n{% endwith %}\n",
+            (3, 2),
+            (3, 5),
+        ),
+        (
+            "{% macro render_items(items) %}\n{% for item in items %}\n{% if item.enabled %}\n{{ item.name }}\n{% endif %}\n{% endfor %}\n{% endmacro %}\n",
+            (3, 2),
+            (3, 5),
+        ),
+    ],
+)
+def test_reporter_translates_loop_backedge_to_no_else_endif_arc_inside_wrapper(
+    tmp_path: Path,
+    source: str,
+    recorded_arc: tuple[int, int],
+    expected_arc: tuple[int, int],
+) -> None:
+    source_file = tmp_path / "template.j2"
+    source_file.write_text(source, encoding="utf-8")
+
+    assert expected_arc in JinjaTemplateFileReporter(str(source_file)).translate_arcs([recorded_arc])
+
+
+def test_reporter_does_not_alias_macro_body_to_enclosing_loop(tmp_path: Path) -> None:
+    source_file = tmp_path / "template.j2"
+    source_file.write_text(
+        "{% for item in items %}\n"
+        "{% macro render_item(item) %}\n"
+        "{% if item.enabled %}\n"
+        "{{ item.name }}\n"
+        "{% endif %}\n"
+        "{% endmacro %}\n"
+        "{{ render_item(item) }}\n"
+        "{% endfor %}\n",
+        encoding="utf-8",
+    )
+
+    assert (3, 5) not in JinjaTemplateFileReporter(str(source_file)).translate_arcs([(3, 1)])
+
+
+def test_reporter_translates_loop_backedge_to_final_no_else_elif_endif_arc(tmp_path: Path) -> None:
+    source_file = tmp_path / "template.j2"
+    source_file.write_text(
+        "{% for item in items %}\n"
+        "{% if item.primary %}\n"
+        "primary {{ item.name }}\n"
+        "{% elif item.secondary %}\n"
+        "secondary {{ item.name }}\n"
+        "{% endif %}\n"
+        "{% endfor %}\n",
+        encoding="utf-8",
+    )
+
+    reporter = JinjaTemplateFileReporter(str(source_file))
+
+    assert (4, 6) in reporter.translate_arcs([(4, 1)])
+    assert (2, 6) not in reporter.translate_arcs([(2, 1)])
+    assert (4, 6) not in reporter.translate_arcs([(4, 5)])
+
+
+def test_reporter_does_not_alias_elif_chain_with_else_to_endif(tmp_path: Path) -> None:
+    source_file = tmp_path / "template.j2"
+    source_file.write_text(
+        "{% for item in items %}\n"
+        "{% if item.primary %}\n"
+        "primary {{ item.name }}\n"
+        "{% elif item.secondary %}\n"
+        "secondary {{ item.name }}\n"
+        "{% else %}\n"
+        "fallback {{ item.name }}\n"
+        "{% endif %}\n"
+        "{% endfor %}\n",
+        encoding="utf-8",
+    )
+
+    assert (4, 8) not in JinjaTemplateFileReporter(str(source_file)).translate_arcs([(4, 1)])
 
 
 def test_reporter_does_not_infer_no_else_endif_arc_from_true_branch(tmp_path: Path) -> None:
@@ -847,115 +941,3 @@ def test_complex_expression_reports_missing_true_output_branch(tmp_path: Path) -
     )
 
     assert analysis.missing_branch_arcs() == {2: [3]}
-
-
-def test_report_does_not_require_compiled_templates(tmp_path: Path) -> None:
-    template_root = tmp_path / "j2templates"
-    compiled_root = template_root / "compiled_templates"
-    source_file = template_root / "simple.j2"
-    coverage_file = tmp_path / ".coverage"
-    coverage_xml = tmp_path / "coverage.xml"
-    script = tmp_path / "render.py"
-    coverage_config = tmp_path / "pyproject.toml"
-
-    compiled_root.mkdir(parents=True)
-    source_file.write_text("{% if enabled %}\nhello\n{% endif %}\n", encoding="utf-8")
-    script.write_text(
-        "from pathlib import Path\n"
-        "from jinja2 import Environment, FileSystemLoader, ModuleLoader\n"
-        "root = Path(__file__).parent / 'j2templates'\n"
-        "compiled = root / 'compiled_templates'\n"
-        "Environment(loader=FileSystemLoader(root)).compile_templates(compiled, zip=None, ignore_errors=False)\n"
-        "Environment(loader=ModuleLoader(compiled)).get_template('simple.j2').render(enabled=True)\n",
-        encoding="utf-8",
-    )
-    coverage_config.write_text(
-        "[tool.coverage.run]\n"
-        "branch = true\n"
-        'core = "ctrace"\n'
-        "parallel = true\n"
-        'plugins = ["coverage_plugins.jinja"]\n'
-        f'source_dirs = ["{template_root.as_posix()}"]\n'
-        "[tool.coverage.coverage_plugins.jinja]\n"
-        f'compiled_template_roots = ["{compiled_root.as_posix()}"]\n',
-        encoding="utf-8",
-    )
-
-    subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "run", "--rcfile", str(coverage_config), "--data-file", str(coverage_file), str(script)],
-        check=True,
-        cwd=tmp_path,
-    )
-    subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "combine", "--rcfile", str(coverage_config), "--data-file", str(coverage_file)],
-        check=True,
-        cwd=tmp_path,
-    )
-    shutil.rmtree(compiled_root)
-
-    subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "report", "--rcfile", str(coverage_config), "--data-file", str(coverage_file), "-m"],
-        check=True,
-        cwd=tmp_path,
-    )
-    subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "xml", "--rcfile", str(coverage_config), "--data-file", str(coverage_file), "-o", str(coverage_xml)],
-        check=True,
-        cwd=tmp_path,
-    )
-
-    xml = coverage_xml.read_text(encoding="utf-8")
-    assert "simple.j2" in xml
-    assert "compiled_templates" not in xml
-
-
-def test_configured_coverage_records_checked_in_templates(tmp_path: Path) -> None:
-    repo_root = Path(__file__).parents[3]
-    coverage_file = tmp_path / ".coverage"
-    script = tmp_path / "render_checked_in_template.py"
-    script.write_text(
-        "from pyavd.constants import EOS_CLI_CONFIG_GEN_JINJA2_PRECOMPILED_TEMPLATE_PATH\n"
-        "from pyavd.templater import Templar\n"
-        "Templar(EOS_CLI_CONFIG_GEN_JINJA2_PRECOMPILED_TEMPLATE_PATH).render_template_from_file(\n"
-        "    'eos/banners.j2',\n"
-        "    {'banners': {'login': 'hello'}},\n"
-        ")\n",
-        encoding="utf-8",
-    )
-
-    env = os.environ.copy()
-    env["COVERAGE_FILE"] = str(coverage_file)
-    env["PYTHONPATH"] = os.pathsep.join(path for path in (str(repo_root), str(repo_root / "python-avd"), env.get("PYTHONPATH", "")) if path)
-    subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "run", "--rcfile=pyproject.toml", str(script)],
-        check=True,
-        cwd=repo_root,
-        env=env,
-    )
-    subprocess.run(
-        [sys.executable, "-m", "coverage", "combine", "--rcfile=pyproject.toml"],
-        check=True,
-        cwd=repo_root,
-        env=env,
-    )
-    coverage_xml = tmp_path / "coverage.xml"
-    xml_result = subprocess.run(  # noqa: S603
-        [sys.executable, "-m", "coverage", "xml", "--rcfile=pyproject.toml", "-o", str(coverage_xml)],
-        check=False,
-        cwd=repo_root,
-        env=env,
-    )
-    assert xml_result.returncode in {0, 2}
-    assert coverage_xml.is_file()
-
-    data = CoverageData(basename=str(coverage_file))
-    data.read()
-    template_file = str(Path("python-avd/pyavd/_eos_cli_config_gen/j2templates/eos/banners.j2").resolve())
-    compiled_file = str(Path("python-avd/pyavd/_eos_cli_config_gen/j2templates/compiled_templates/eos__banners.py").resolve())
-
-    assert template_file in data.measured_files()
-    assert data.file_tracer(template_file) == "coverage_plugins.jinja.JinjaTemplateCoveragePlugin"
-    assert data.lines(template_file)
-    assert compiled_file not in data.measured_files()
-
-    assert 'filename="python-avd/pyavd/_eos_cli_config_gen/j2templates/eos/banners.j2"' in coverage_xml.read_text(encoding="utf-8")
